overhauled writers
