+++ src/OpenRGB.NET/ISpanReader.cs
@@ -4,5 +4,5 @@ namespace OpenRGB.NET;
 
 internal interface ISpanReader<out T>
 {
-    T ReadFrom(ref SpanReader reader, ProtocolVersion protocolVersion = default, int index = default);
+    T ReadFrom(ref SpanReader reader, ProtocolVersion? protocolVersion = default, int? index = default, int? outerCount = default);
 }
