+++ src/OpenRGB.NET/Models/Color.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using OpenRGB.NET.Utils;
+using System.Runtime.InteropServices;
 
 namespace OpenRGB.NET;
 
@@ -9,6 +8,7 @@ namespace OpenRGB.NET;
 /// <param name="R">The Red component</param>
 /// <param name="G">The Green component</param>
 /// <param name="B">The Blue component</param>
+[StructLayout(LayoutKind.Sequential, Pack = 1)]
 public readonly record struct Color(byte R = 0, byte G = 0, byte B = 0)
 {
     /// <summary>
@@ -28,38 +28,4 @@ public readonly record struct Color(byte R = 0, byte G = 0, byte B = 0)
 
     //Added for padding, so Color fits within 4 bytes.
     private readonly byte UnusedAlpha = 0;
-
-    private static Color ReadFrom(ref SpanReader reader)
-    {
-        var r = reader.Read<byte>();
-        var g = reader.Read<byte>();
-        var b = reader.Read<byte>();
-        var a = reader.Read<byte>();
-
-        return new Color(r, g, b);
-    }
-
-    internal static Color[] ReadManyFrom(ref SpanReader reader, int count)
-    {
-        if (count < 0)
-            throw new ArgumentOutOfRangeException(nameof(count), "Count must be greater than or equal to 0.");
-
-        if (count == 0)
-            return [];
-        
-        var colors = new Color[count];
-
-        for (var i = 0; i < count; i++)
-            colors[i] = ReadFrom(ref reader);
-
-        return colors;
-    }
-
-    internal void WriteTo(ref SpanWriter writer)
-    {
-        writer.Write(R);
-        writer.Write(G);
-        writer.Write(B);
-        writer.Write(UnusedAlpha);
-    }
 }
