+++ src/OpenRGB.NET/Models/Led.cs
@@ -1,5 +1,3 @@
-using OpenRGB.NET.Utils;
-
 namespace OpenRGB.NET;
 
 /// <summary>
@@ -7,7 +5,7 @@ namespace OpenRGB.NET;
 /// </summary>
 public class Led
 {
-    private Led(int index, string name, uint value)
+    internal Led(int index, string name, uint value)
     {
         Index = index;
         Name = name;
@@ -28,30 +26,6 @@ public class Led
     ///     Device specific led value. Most likely not useful for the clients.
     /// </summary>
     public uint Value { get;  }
-
-    private static Led ReadFrom(ref SpanReader reader, int index)
-    {
-        var name = reader.ReadLengthAndString();
-        var value = reader.Read<uint>();
-        
-        return new Led(index, name, value);
-    }
-
-    /// <summary>
-    ///     Decodes a byte array into a LED array.
-    ///     Increments the offset accordingly.
-    /// </summary>
-    /// <param name="reader"></param>
-    /// <param name="ledCount"></param>
-    internal static Led[] ReadManyFrom(ref SpanReader reader, ushort ledCount)
-    {
-        var leds = new Led[ledCount];
-
-        for (var i = 0; i < ledCount; i++)
-            leds[i] = ReadFrom(ref reader, i);
-
-        return leds;
-    }
 
     /// <inheritdoc />
     public override string ToString()
