+++ src/OpenRGB.NET/Models/Segment.cs
@@ -1,5 +1,3 @@
-using OpenRGB.NET.Utils;
-
 namespace OpenRGB.NET;
 
 /// <summary>
@@ -7,7 +5,7 @@ namespace OpenRGB.NET;
 /// </summary>
 public class Segment
 {
-    private Segment(int index, string name, ZoneType type, uint start, uint ledCount)
+    internal Segment(int index, string name, ZoneType type, uint start, uint ledCount)
     {
         Index = index;
         Name = name;
@@ -40,24 +38,4 @@ public class Segment
     ///     The number of LEDs in the segment.
     /// </summary>
     public uint LedCount { get; }
-
-    private static Segment ReadFrom(ref SpanReader reader, int index)
-    {
-        var name = reader.ReadLengthAndString();
-        var type = (ZoneType)reader.Read<uint>();
-        var start = reader.Read<uint>();
-        var ledCount = reader.Read<uint>();
-
-        return new Segment(index, name, type, start, ledCount);
-    }
-
-    internal static Segment[] ReadManyFrom(ref SpanReader reader, ushort segmentCount)
-    {
-        var segments = new Segment[segmentCount];
-
-        for (var i = 0; i < segmentCount; i++)
-            segments[i] = ReadFrom(ref reader, i);
-
-        return segments;
-    }
 }
